Add new config value: review_keep_comments
In Re:VIEW, comments are not allowed everywhere.  For example, it
is hard to write comments inside table, lists and so on.

This option make conversion more easiler and safer. I believe
users of this extensions don't want to modify Re:VIEW articles. So
I think this does not bring any problem for major case.